LayoutTests/http/tests/loading/resourceLoadStatistics/request-storage-access-top-frame-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
Tests that top frame storage access is always granted in case the page requests it.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS requestStorageAccessReturnValue is true
PASS document.hasStorageAccess is true
PASS requestStorageAccessReturnValue is true
PASS document.hasStorageAccess is true
PASS successfullyParsed is true

TEST COMPLETE
Request Access
